+++ src/countries/us.ts
@@ -9,6 +9,8 @@ export const us: Country = {
 		sv: 'Amerikas förenta stater',
 		da: 'Amerikas Forenede Stater',
 		sq: 'Shtetet e Bashkuara të Amerikës',
+		pt: 'Estados Unidos da América',
+		tr: 'Amerika Birleşik Devletleri',
 		xa: 'The Star-Spangled Banner',
 	},
 	flag: '🇺🇸',